chore(core): delete unused functions
