+++ core/src/apps/monero/xmr/bulletproof.py
@@ -203,15 +203,6 @@ def _sc_mul(dst: bytearray | None, a: bytes, b: bytes | crypto.Scalar) -> bytear
     return dst
 
 
-def _sc_mul8(dst: bytearray | None, a: bytes) -> bytearray:
-    dst = _ensure_dst_key(dst)
-    crypto.decodeint_into_noreduce(_tmp_sc_1, a)
-    crypto.decodeint_into_noreduce(_tmp_sc_2, _EIGHT)
-    crypto.sc_mul_into(_tmp_sc_3, _tmp_sc_1, _tmp_sc_2)
-    crypto.encodeint_into(dst, _tmp_sc_3)
-    return dst
-
-
 def _sc_muladd(
     dst: ScalarDst | None,
     a: bytes | crypto.Scalar,
@@ -455,11 +446,6 @@ class KeyVBase(Generic[T]):
     def read(self, idx: int, buff: bytes, offset: int = 0) -> bytes:
         raise NotImplementedError
 
-    def slice(self, res, start: int, stop: int):
-        for i in range(start, stop):
-            res[i - start] = self[i]
-        return res
-
     def slice_view(self, start: int, stop: int) -> "KeyVSliced":
         return KeyVSliced(self, start, stop)
 
@@ -634,32 +620,6 @@ class KeyV(KeyVBaseType[T]):
         self.size = nsize
         self._set_mv()
 
-    def realloc_init_from(self, nsize, src, offset: int = 0, collect: int = False):
-        if not isinstance(src, KeyV):
-            raise ValueError("KeyV supported only")
-        self.realloc(nsize, collect)
-
-        if not self.chunked and not src.chunked:
-            assert isinstance(self.d, bytearray)
-            assert isinstance(src.d, (bytes, bytearray))
-            memcpy(self.d, 0, src.d, offset << 5, nsize << 5)
-
-        elif self.chunked and not src.chunked or self.chunked and src.chunked:
-            for i in range(nsize):
-                self.read(i, src.to(i + offset))
-
-        elif not self.chunked and src.chunked:
-            assert isinstance(self.d, bytearray)
-            assert isinstance(src.d, list)
-            for i in range(nsize >> _CHBITS):
-                memcpy(
-                    self.d,
-                    i << 11,
-                    src.d[i + (offset >> _CHBITS)],
-                    (offset & (_CHSIZE - 1)) << 5 if i == 0 else 0,
-                    nsize << 5 if i <= nsize >> _CHBITS else (nsize & _CHSIZE) << 5,
-                )
-
 
 class KeyVEval(KeyVBase):
     """
@@ -696,40 +656,6 @@ class KeyVEval(KeyVBase):
         else:
             memcpy(buff, offset, self.buff, 0, 32)
         return buff if buff else self.buff
-
-
-class KeyVSized(KeyVBase):
-    """
-    Resized vector, wrapping possibly larger vector
-    (e.g., precomputed, but has to have exact size for further computations)
-    """
-
-    __slots__ = ("current_idx", "size", "wrapped")
-
-    def __init__(self, wrapped, new_size):
-        super().__init__(new_size)
-        self.wrapped = wrapped
-
-    def __getitem__(self, item):
-        return self.wrapped[self.idxize(item)]
-
-    def __setitem__(self, key, value):
-        self.wrapped[self.idxize(key)] = value
-
-
-class KeyVConst(KeyVBase):
-    __slots__ = ("current_idx", "size", "elem")
-
-    def __init__(self, size, elem, copy=True):
-        super().__init__(size)
-        self.elem = _init_key(elem) if copy else elem
-
-    def __getitem__(self, item):
-        return self.elem
-
-    def to(self, idx: int, buff: bytearray, offset: int = 0):
-        memcpy(buff, offset, self.elem, 0, 32)
-        return buff if buff else self.elem
 
 
 class KeyVPrecomp(KeyVBase):
@@ -1307,10 +1233,6 @@ def _ensure_dst_keyvect(dst=None, size: int | None = None):
     return dst
 
 
-def _const_vector(val, elems=_BP_N, copy: bool = True) -> KeyVConst:
-    return KeyVConst(elems, val, copy)
-
-
 def _vector_exponent_custom(A, B, a, b, dst=None, a_raw=None, b_raw=None):
     """
     \\sum_{i=0}^{|A|}  a_i A_i + b_i B_i
@@ -1498,16 +1420,6 @@ def _cross_inner_product(l0, r0, l1, r1):
         _gc_iter(i)
 
     return crypto_helpers.encodeint(sc_t1), crypto_helpers.encodeint(sc_t2)
-
-
-def _vector_gen(dst, size, op):
-    dst = _ensure_dst_keyvect(dst, size)
-    for i in range(size):
-        dst.to(i, _tmp_bf_0)
-        op(i, _tmp_bf_0)
-        dst.read(i, _tmp_bf_0)
-        _gc_iter(i)
-    return dst
 
 
 def _vector_dup(x, n, dst=None):
@@ -2390,9 +2302,6 @@ class BulletProofPlusBuilder:
         return KeyVPrecomp(
             size, self.Hprec, lambda i, d: _get_exponent_plus(d, _XMR_H, i * 2)
         )
-
-    def vector_exponent(self, a, b, dst=None, a_raw=None, b_raw=None):
-        return _vector_exponent_custom(self.Gprec, self.Hprec, a, b, dst, a_raw, b_raw)
 
     def prove(
         self, sv: list[crypto.Scalar], gamma: list[crypto.Scalar]
