+++ core/src/trezor/ui/components/tt/scroll.py
@@ -5,8 +5,8 @@ from trezor import loop, res, ui, utils, wire, workflow
 from trezor.enums import ButtonRequestType
 from trezor.messages import ButtonAck, ButtonRequest
 
-from ..common.confirm import CANCELLED, CONFIRMED, GO_BACK, SHOW_PAGINATED
-from .button import Button, ButtonCancel, ButtonConfirm, ButtonDefault
+from ..common.confirm import GO_BACK, SHOW_PAGINATED
+from .button import Button, ButtonDefault
 from .confirm import Confirm
 from .swipe import SWIPE_DOWN, SWIPE_UP, SWIPE_VERTICAL, Swipe
 from .text import (
@@ -182,168 +182,6 @@ class AskPaginated(ui.Component):
 
         def read_content(self) -> list[str]:
             return self.content.read_content()
-
-
-class PageWithButtons(ui.Component):
-    def __init__(
-        self,
-        content: ui.Component,
-        paginated: "PaginatedWithButtons",
-        index: int,
-        count: int,
-    ) -> None:
-        super().__init__()
-        self.content = content
-        self.paginated = paginated
-        self.index = index
-        self.count = count
-
-        # somewhere in the middle, we can go up or down
-        left = res.load(ui.ICON_BACK)
-        left_style = ButtonDefault
-        right = res.load(ui.ICON_CLICK)
-        right_style = ButtonDefault
-
-        if self.index == 0:
-            # first page, we can cancel or go down
-            left = res.load(ui.ICON_CANCEL)
-            left_style = ButtonCancel
-            right = res.load(ui.ICON_CLICK)
-            right_style = ButtonDefault
-        elif self.index == count - 1:
-            # last page, we can go up or confirm
-            left = res.load(ui.ICON_BACK)
-            left_style = ButtonDefault
-            right = res.load(ui.ICON_CONFIRM)
-            right_style = ButtonConfirm
-
-        self.left = Button(ui.grid(8, n_x=2), left, left_style)
-        self.left.on_click = self.on_left
-
-        self.right = Button(ui.grid(9, n_x=2), right, right_style)
-        self.right.on_click = self.on_right
-
-    def dispatch(self, event: int, x: int, y: int) -> None:
-        self.content.dispatch(event, x, y)
-        self.left.dispatch(event, x, y)
-        self.right.dispatch(event, x, y)
-
-    def on_left(self) -> None:
-        if self.index == 0:
-            self.paginated.on_cancel()
-        else:
-            self.paginated.on_up()
-
-    def on_right(self) -> None:
-        if self.index == self.count - 1:
-            self.paginated.on_confirm()
-        else:
-            self.paginated.on_down()
-
-    if __debug__:
-
-        def read_content(self) -> list[str]:
-            return self.content.read_content()
-
-
-class PaginatedWithButtons(ui.Layout):
-    def __init__(
-        self, pages: list[ui.Component], page: int = 0, one_by_one: bool = False
-    ) -> None:
-        super().__init__()
-        self.pages = [
-            PageWithButtons(p, self, i, len(pages)) for i, p in enumerate(pages)
-        ]
-        self.page = page
-        self.one_by_one = one_by_one
-
-    def dispatch(self, event: int, x: int, y: int) -> None:
-        pages = self.pages
-        page = self.page
-        pages[page].dispatch(event, x, y)
-        if event is ui.RENDER:
-            render_scrollbar(len(pages), page)
-
-    def on_up(self) -> None:
-        self.page = max(self.page - 1, 0)
-        self.pages[self.page].dispatch(ui.REPAINT, 0, 0)
-        self.on_change()
-
-    def on_down(self) -> None:
-        self.page = min(self.page + 1, len(self.pages) - 1)
-        self.pages[self.page].dispatch(ui.REPAINT, 0, 0)
-        self.on_change()
-
-    def on_confirm(self) -> None:
-        raise ui.Result(CONFIRMED)
-
-    def on_cancel(self) -> None:
-        raise ui.Result(CANCELLED)
-
-    def on_change(self) -> None:
-        if self.one_by_one:
-            raise ui.Result(self.page)
-
-    if __debug__:
-
-        def read_content(self) -> list[str]:
-            return self.pages[self.page].read_content()
-
-        def create_tasks(self) -> tuple[loop.AwaitableTask, ...]:
-            from apps.debug import confirm_signal
-
-            return super().create_tasks() + (confirm_signal(),)
-
-
-def paginate_text(
-    text: str,
-    header: str,
-    font: int = ui.NORMAL,
-    header_icon: str = ui.ICON_DEFAULT,
-    icon_color: int = ui.ORANGE_ICON,
-    break_words: bool = False,
-    confirm: Callable[[ui.Component], ui.Layout] = Confirm,
-) -> ui.Layout:
-    span = Span(text, 0, font, break_words=break_words)
-    if span.count_lines() <= TEXT_MAX_LINES:
-        result = Text(
-            header,
-            header_icon=header_icon,
-            icon_color=icon_color,
-            new_lines=False,
-            break_words=break_words,
-        )
-        result.content = [font, text]
-        return confirm(result)
-
-    else:
-        pages: list[ui.Component] = []
-        span.reset(
-            text, 0, font, break_words=break_words, line_width=LINE_WIDTH_PAGINATED
-        )
-        while span.has_more_content():
-            # advance to first line of the page
-            span.next_line()
-            page = Text(
-                header,
-                header_icon=header_icon,
-                icon_color=icon_color,
-                new_lines=False,
-                content_offset=0,
-                char_offset=span.start,
-                line_width=LINE_WIDTH_PAGINATED,
-                break_words=break_words,
-                render_page_overflow=False,
-            )
-            page.content = [font, text]
-            pages.append(page)
-
-            # roll over the remaining lines on the page
-            for _ in range(TEXT_MAX_LINES - 1):
-                span.next_line()
-
-        pages[-1] = confirm(pages[-1])
-        return Paginated(pages)
 
 
 PAGEBREAK = 0, ""
