+++ core/tests/test_apps.common.paths.py
@@ -2,6 +2,13 @@ from common import *
 from trezor.utils import ensure
 from apps.common.paths import *
 
+# NOTE: moved into tests not to occupy flash space
+# in firmware binary, when it is not used in production
+class NeverMatchingSchema:
+    @staticmethod
+    def match(path: "Bip32Path") -> bool:
+        return False
+
 
 class TestPaths(unittest.TestCase):
     def test_is_hardened(self):
